+++ apps/console/src/features/core/utils/route-utils.ts
@@ -401,7 +401,7 @@ export class RouteUtils {
             },
             {
                 category: settings,       
-                id: "remoteLogIngest"
+                id: "remoteLogging"
             },
             {
                 category: settings,